+++ packages/lit-dev-content/samples/examples/properties-has-changed/my-element.ts
@@ -1,37 +1,7 @@
-import { LitElement, html, PropertyValues} from "lit";
-import {customElement, property, query} from 'lit/decorators.js';
-import {isSameDate, localDateFromUTC} from './date-utils.js';
-
-@customElement('date-display')
-class DateDisplay extends LitElement {
-  @property({
-    // For wrapper types like Date, simple comparison won't work, because
-    // each Date is a new object
-    hasChanged: (value?: Date, oldValue?: Date) => {
-      return !isSameDate(value, oldValue);
-    }
-  })
-  date?: Date;
-
-  @query('#datefield')
-  datefield!: HTMLSpanElement;
-
-  frames = [
-    {backgroundColor: '#fff'},
-    {backgroundColor: '#324fff' },
-    {backgroundColor: '#fff'}
-  ];
-
-  render() {
-    return html`<span id="datefield">${this.date?.toLocaleDateString()}</span>`;
-  }
-
-  updated(changed: PropertyValues) {
-    if (changed.has('date')) {
-      this.datefield.animate(this.frames, 1000);
-    }
-  }
-}
+import { LitElement, html} from "lit";
+import {customElement, property} from 'lit/decorators.js';
+import {localDateFromUTC} from './date-utils.js';
+import './date-display.js';
 
 @customElement('my-element')
 class MyElement extends LitElement {
