+++ packages/lit-dev-content/samples/examples/properties-has-changed/project.json
@@ -4,6 +4,7 @@
   "description": "Customizing change detection.",
   "section": "Reactive properties",
   "files": {
+    "date-display.ts": {},
     "my-element.ts": {},
     "date-utils.ts": {},
     "index.html": {}
